LOOPCARE-544: Update registration model

See merge request mobidev/loopcare-frontend!99

+++ lib/features/authentication/application/dto/sign_up_data.dart
@@ -12,9 +12,9 @@ class SignUpData {
   final bool isConsentApproved;
   final bool isLegalApproved;
 
-  final int height;
-  final int weight;
-  final int bmi;
+  final double height;
+  final double weight;
+  final double bmi;
   final String birthDate;
   final String gender;
   final String bioGender;
